move gzip logic into new class ArchivableFile
